Final minor conditional tweak for ELN
Yesterday, my pull request was rushed.  I only waited two hours through the build before thinking it was fixed.
Because I rushed I missed one other conditional, and the builds still don't builds in ELN.
This time, with this pull request, I waited the full 8+ hours.  It builds successfully, completely.
We are still in a bit of a time crunch, so if you wouldn't mind merging this pull request, I would appreciate it.

Signed-off-by: Troy Dawson <tdawson@redhat.com>